backend/ai-microservice/requirements.txt
fastapi
uvicorn
google-genai
pydantic
dotenv